requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
scikit-learn==1.4.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
